test(a2r): update expected outputs for transpiler changes
Regenerated all a2r test expected.rs files to match updated transpiler output
(json.get returning &Value, removed .cloned(), sync http.post_sync, etc.)

+++ crates/auto-lang/test/a2r/09_option_result/027_list_methods/list_methods.expected.rs
@@ -9,7 +9,7 @@ fn main() {
     let list = List::new();
 
 
-    let len = list.len();
+    let len = (list.len() as i32);
     let is_empty = list.is_empty();
 
 
@@ -17,5 +17,5 @@ fn main() {
     list.push(20);
     list.push(30);
 
-    let new_len = list.len();
+    let new_len = (list.len() as i32);
 }
